task_management_app/backend/requirements.txt
fastapi
uvicorn
pydantic
passlib
bcrypt
PyJWT
python-multipart
